pyproject.toml
[project]
name = "ask-pesu-subreddit-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "httpx>=0.28.1",
    "praw>=7.8.1",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "secret_required: marks tests that require secrets or environment variables (e.g. TEST_PRN, TEST_PASSWORD)"
]

[tool.ruff.lint]
select = [
    "E", # Pycodestyle errors
    "F", # Pyflakes errors
    "W", # Pycodestyle warnings
    "UP", # Pyupgrade rules
    "I", # Import related rules
    "C90", # Complexity rules
    "D", # Documentation rules
    "ANN", # Annotations rules
    "TYP", # Type checking rules
    "N", # Naming rules
    "Q000", # Quotation style rules
    "RET", # Return type rules
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = [
    "tests/"
]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
